pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-mdfile-generator"
version = "0.1.0"
description = "Lightweight TUI to generate claude.md files from reusable modules"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "questionary>=2.0",
    "rich>=13.0",
]

[project.scripts]
claude-md = "claude_mdfile_generator.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
]
